DeepLearningTemplate/requirements.txt
numpy==1.22.0
pandas==1.3.4
Pillow==9.0.1
pytorch_lightning==1.4.9
ray[tune]==1.9.1
ruamel.yaml==0.17.17
scikit_learn==1.0.2
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tqdm==4.62.3
